move usePrevious it's a dependency
