+++ webui/react/src/components/TableFilterDropdown.tsx
@@ -3,8 +3,9 @@ import { FilterDropdownProps } from 'antd/es/table/interface';
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { FixedSizeList, ListChildComponentProps } from 'react-window';
 
-import usePrevious from 'hooks/usePrevious';
-import Icon from 'shared/components/Icon/Icon';
+import usePrevious from 'shared/hooks/usePrevious';
+
+import Icon from '../shared/components/Icon/Icon';
 
 import css from './TableFilterDropdown.module.scss';
